+++ InspireHome/src/main/java/com/inspire/startup/security/JwtAuthenticationFilter.java
@@ -46,7 +46,7 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 	            String jwt = getJwtFromRequest(request);
 
 	            if (StringUtils.hasText(jwt) && tokenProvider.validateToken(jwt)) {
-	                Long userId = tokenProvider.getUserIdFromJWT(jwt);
+	                Integer userId = tokenProvider.getUserIdFromJWT(jwt);
 
 	                /*
 	                    Note that you could also encode the user's username and roles inside JWT claims
